# docs/api/radiance-openapi.json
{
  "components": {
    "schemas": {
...
      "AssemblyPhotometricBoundsResponse": {
        "additionalProperties": false,
        "properties": {
          "x_max": {
            "title": "X Max",
            "type": "number"
          },
          "x_min": {
            "title": "X Min",
            "type": "number"
          },
          "y_max": {
            "title": "Y Max",
            "type": "number"
          },
          "y_min": {
            "title": "Y Min",
            "type": "number"
          },
          "z_m": {
            "title": "Z M",
            "type": "number"
          }
        },
        "required": [
          "x_min",
          "x_max",
          "y_min",
          "y_max",
          "z_m"
        ],
        "title": "AssemblyPhotometricBoundsResponse",
        "type": "object"
      },
      "AssemblyPhotometricColorScaleResponse": {
        "additionalProperties": false,
        "properties": {
          "clamp": {
            "const": true,
            "title": "Clamp",
            "type": "boolean"
          },
          "source": {
            "const": "simulation-visualization-mean-ppfd",
            "title": "Source",
            "type": "string"
          },
          "vmax": {
            "title": "Vmax",
            "type": "number"
          },
          "vmin": {
            "title": "Vmin",
            "type": "number"
          }
        },
        "required": [
          "vmin",
          "vmax",
          "source",
          "clamp"
        ],
        "title": "AssemblyPhotometricColorScaleResponse",
        "type": "object"
      },
      "AssemblyPhotometricColormapResponse": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "const": "viridis",
            "title": "Name",
            "type": "string"
          },
          "normalization": {
            "const": "linear-clamped",
            "title": "Normalization",
            "type": "string"
          },
          "source": {
            "const": "matplotlib",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "name",
          "source",
          "normalization"
        ],
        "title": "AssemblyPhotometricColormapResponse",
        "type": "object"
      },
      "AssemblyPhotometricLayerResponse": {
        "additionalProperties": false,
        "properties": {
          "bounds_m": {
            "$ref": "#/components/schemas/AssemblyPhotometricBoundsResponse"
          },
          "color_scale": {
            "$ref": "#/components/schemas/AssemblyPhotometricColorScaleResponse"
          },
          "colormap": {
            "$ref": "#/components/schemas/AssemblyPhotometricColormapResponse"
          },
          "encoding": {
            "const": "float32-le",
            "title": "Encoding",
            "type": "string"
          },
          "grid_height": {
            "title": "Grid Height",
            "type": "integer"
          },
          "grid_width": {
            "title": "Grid Width",
            "type": "integer"
          },
          "max_ppfd": {
            "title": "Max Ppfd",
            "type": "number"
          },
          "mean_ppfd": {
            "title": "Mean Ppfd",
            "type": "number"
          },
          "min_ppfd": {
            "title": "Min Ppfd",
            "type": "number"
          },
          "mode": {
            "enum": [
              "SMD",
              "Competitor",
              "1000W HPS"
            ],
            "title": "Mode",
            "type": "string"
          },
          "mode_label": {
            "title": "Mode Label",
            "type": "string"
          },
          "orientation": {
            "$ref": "#/components/schemas/AssemblyPhotometricOrientationResponse"
          },
          "schema_version": {
            "const": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "target_ppfd": {
            "title": "Target Ppfd",
            "type": "number"
          },
          "units": {
            "const": "\u00b5mol/m\u00b2/s",
            "title": "Units",
            "type": "string"
          },
          "value_count": {
            "title": "Value Count",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "mode",
          "mode_label",
          "units",
          "grid_width",
          "grid_height",
          "value_count",
          "bounds_m",
          "min_ppfd",
          "max_ppfd",
          "mean_ppfd",
          "target_ppfd",
          "encoding",
          "colormap",
          "color_scale",
          "orientation",
          "warnings"
        ],
        "title": "AssemblyPhotometricLayerResponse",
        "type": "object"
      },
      "AssemblyPhotometricOrientationResponse": {
        "additionalProperties": false,
        "properties": {
          "column_axis": {
            "const": "x",
            "title": "Column Axis",
            "type": "string"
          },
          "column_order": {
            "const": "ascending",
            "title": "Column Order",
            "type": "string"
          },
          "plane": {
            "const": "xy",
            "title": "Plane",
            "type": "string"
          },
          "row_axis": {
            "const": "y",
            "title": "Row Axis",
            "type": "string"
          },
          "row_order": {
            "const": "ascending",
            "title": "Row Order",
            "type": "string"
          },
          "storage_order": {
            "const": "row-major",
            "title": "Storage Order",
            "type": "string"
          },
          "value_index": {
            "const": "row * grid_width + column",
            "title": "Value Index",
            "type": "string"
          }
        },
        "required": [
          "plane",
          "column_axis",
          "row_axis",
          "column_order",
          "row_order",
          "storage_order",
          "value_index"
        ],
        "title": "AssemblyPhotometricOrientationResponse",
        "type": "object"
      },
...
      "PublicErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "const": false,
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "error",
          "message",
          "correlation_id"
        ],
        "title": "PublicErrorResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Horticulture Lighting Simulator API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/radiance/assembly-photometric-layer": {
      "get": {
        "operationId": "get_radiance_assembly_photometric_layer",
        "parameters": [
          {
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "default": "SMD",
              "title": "Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "execution_mode",
            "required": false,
            "schema": {
              "default": "precomputed",
              "title": "Execution Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sim_mode",
            "required": false,
            "schema": {
              "default": "standard",
              "title": "Sim Mode",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "target_ppfd",
            "required": false,
            "schema": {
              "default": 1000.0,
              "title": "Target Ppfd",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "peak_capping_enabled",
            "required": false,
            "schema": {
              "default": false,
              "title": "Peak Capping Enabled",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "match_system_ppe",
            "required": false,
            "schema": {
              "default": false,
              "title": "Match System Ppe",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "length_ft",
            "required": false,
            "schema": {
              "default": 10.0,
              "title": "Length Ft",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "width_ft",
            "required": false,
            "schema": {
              "default": 10.0,
              "title": "Width Ft",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "w_min",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "W Min"
            }
          },
          {
            "in": "query",
            "name": "w_max",
            "required": false,
            "schema": {
              "default": 100.0,
              "title": "W Max",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "competitor_layout",
            "required": false,
            "schema": {
              "default": "full",
              "title": "Competitor Layout",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hps_coverage_ft",
            "required": false,
            "schema": {
              "default": 4.0,
              "title": "Hps Coverage Ft",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "hps_ies_variant",
            "required": false,
            "schema": {
              "default": "karma",
              "title": "Hps Ies Variant",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mount_z_m",
            "required": false,
            "schema": {
              "default": 0.4572,
              "title": "Mount Z M",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "sp_z_m",
            "required": false,
            "schema": {
              "default": 0.4572,
              "title": "Sp Z M",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "hps_z_m",
            "required": false,
            "schema": {
              "default": 0.4572,
              "title": "Hps Z M",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "basis_backend",
            "required": false,
            "schema": {
              "default": "rtrace",
              "title": "Basis Backend",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "plants_enabled",
            "required": false,
            "schema": {
              "default": false,
              "title": "Plants Enabled",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "plant_seed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Seed"
            }
          },
          {
            "in": "query",
            "name": "plant_rows",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Rows"
            }
          },
          {
            "in": "query",
            "name": "plant_columns",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Columns"
            }
          },
          {
            "in": "query",
            "name": "plant_spacing_m",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Spacing M"
            }
          },
          {
            "in": "query",
            "name": "plant_height_m",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Height M"
            }
          },
          {
            "in": "query",
            "name": "plant_canopy_radius_m",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Canopy Radius M"
            }
          },
          {
            "in": "query",
            "name": "plant_leaf_count",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Leaf Count"
            }
          },
          {
            "in": "query",
            "name": "plant_growth_stage",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plant Growth Stage"
            }
          },
          {
            "in": "query",
            "name": "fspm_target_ppfd_umol_m2_s",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fspm Target Ppfd Umol M2 S"
            }
          },
          {
            "in": "query",
            "name": "fspm_target_tolerance_umol_m2_s",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fspm Target Tolerance Umol M2 S"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "artifact_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Artifact Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssemblyPhotometricLayerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Request Entity Too Large"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Radiance Assembly Photometric Layer"
      }
    },
...
  }
}